perbaikan view, penambahan role aplikasi pindai & penambahan icon

+++ app/Http/Controllers/LaporanPenilaianController.php
@@ -36,15 +36,19 @@ class LaporanPenilaianController extends Controller
      */
     public function store(Request $request)
     {
-        $key=pemberitahuanPenilaian::all()->find($request->pemberitahuan_penilaian_id);
-        if (!isset($key->penyampaianLaporan)) {
-            $ValidatedData=$request->validate([
-                'nomorLaporan'=>'required',
-                'tanggalLaporan'=>'required',
-                'pemberitahuan_penilaian_id'=>'required',
-            ]);
-            laporanPenilaian::create($ValidatedData);
-            return redirect('/penilaian/'. $key->permohonan_penilaian_id);
+        if (auth()->user()->jabatan === '01' || auth()->user()->jabatan === '02' || auth()->user()->jabatan === '09' || auth()->user()->jabatan === '10' || auth()->user()->jabatan === '11') {
+            $key=pemberitahuanPenilaian::all()->find($request->pemberitahuan_penilaian_id);
+            if (!isset($key->penyampaianLaporan)) {
+                $ValidatedData=$request->validate([
+                    'nomorLaporan'=>'required',
+                    'tanggalLaporan'=>'required',
+                    'pemberitahuan_penilaian_id'=>'required',
+                ]);
+                laporanPenilaian::create($ValidatedData);
+                return redirect('/penilaian/'. $key->permohonan_penilaian_id);
+            }else{
+                abort(403);
+            }
         }else{
             abort(403);
         }
@@ -58,9 +62,11 @@ class LaporanPenilaianController extends Controller
      */
     public function show(laporanPenilaian $laporanpenilaian)
     {
-        //
-
-        return json_encode($laporanpenilaian->barang);
+        if (auth()->user()->jabatan === '01' || auth()->user()->jabatan === '02' || auth()->user()->jabatan === '09' || auth()->user()->jabatan === '10' || auth()->user()->jabatan === '11') {
+            return json_encode($laporanpenilaian->barang);
+        }else{
+            abort(403);
+        }
     }
 
     /**
@@ -94,12 +100,16 @@ class LaporanPenilaianController extends Controller
      */
     public function destroy(laporanPenilaian $laporanpenilaian)
     {
-        $key = $laporanpenilaian->pemberitahuanPenilaian->penyampaianLaporan;
-        $key2 = $laporanpenilaian->barang->first();
-        if(!isset($key2)){
-            if (!isset($key)) {
-                $laporanpenilaian->delete();
-                return redirect('/penilaian/'. $laporanpenilaian->pemberitahuanPenilaian->permohonan_penilaian_id);
+        if (auth()->user()->jabatan === '01' || auth()->user()->jabatan === '02' || auth()->user()->jabatan === '09' || auth()->user()->jabatan === '10' || auth()->user()->jabatan === '11') {
+            $key = $laporanpenilaian->pemberitahuanPenilaian->penyampaianLaporan;
+            $key2 = $laporanpenilaian->barang->first();
+            if(!isset($key2)){
+                if (!isset($key)) {
+                    $laporanpenilaian->delete();
+                    return redirect('/penilaian/'. $laporanpenilaian->pemberitahuanPenilaian->permohonan_penilaian_id);
+                }else{
+                    abort(403);
+                }
             }else{
                 abort(403);
             }
