+++ app/Http/Controllers/PermohonanController.php
@@ -16,10 +16,14 @@ class PermohonanController extends Controller
      */
     public function index()
     {
-        //
-        return view('pindaiPermohonan',[
-            'data'=>permohonan::orderBy('created_at', 'desc')->get(),
-        ]);
+        if (auth()->user()->jabatan === '01' || auth()->user()->jabatan === '03' || auth()->user()->jabatan === '12') {
+            return view('pindai.Permohonan',[
+                'data'=>permohonan::orderBy('created_at', 'desc')->get(),
+                'permohonanview'=>''
+            ]);
+        }else{
+            abort(403);
+        }
     }
 
     /**
@@ -40,45 +44,48 @@ class PermohonanController extends Controller
      */
     public function store(Request $request)
     {
-        //
-        $ValidatedData=$request->validate([
-            'nomorSurat'=>'required',
-            'hal'=>'required',
-            'tanggalSurat'=>'required',
-            'pemohon'=>'required|numeric|digits:6',
-            'tanggalDiTerima'=>'required',
-        ]);
-        $tiket_id=tiket::count();
-        $tiket_id++;
-        $tiket='PKN';
-        if ($tiket_id < 10) {
-            $tiket .= '0000';
-        }elseif ($tiket_id < 100) {
-            $tiket .= '000';
-        }elseif ($tiket_id < 1000) {
-            $tiket .= '000';
-        }elseif ($tiket_id < 10000) {
-            $tiket .= '00';
-        }elseif ($tiket_id < 100000) {
-            $tiket .= '0';
-        }elseif ($tiket_id >= 100000) {
-            $tiket .= '';
+        if (auth()->user()->jabatan === '01' || auth()->user()->jabatan === '03' || auth()->user()->jabatan === '12') {
+            $ValidatedData=$request->validate([
+                'nomorSurat'=>'required',
+                'hal'=>'required',
+                'tanggalSurat'=>'required',
+                'pemohon'=>'required|numeric|digits:6',
+                'tanggalDiTerima'=>'required',
+            ]);
+            $tiket_id=tiket::count();
+            $tiket_id++;
+            $tiket='PKN';
+            if ($tiket_id < 10) {
+                $tiket .= '0000';
+            }elseif ($tiket_id < 100) {
+                $tiket .= '000';
+            }elseif ($tiket_id < 1000) {
+                $tiket .= '000';
+            }elseif ($tiket_id < 10000) {
+                $tiket .= '00';
+            }elseif ($tiket_id < 100000) {
+                $tiket .= '0';
+            }elseif ($tiket_id >= 100000) {
+                $tiket .= '';
+            }
+            $tiket .= $tiket_id; 
+            $tiketbaru = tiket::create([
+                'tiket'=>$tiket,
+                'permohonan'=>1,
+            ]);
+                
+            permohonan::create([
+                'nomorSurat'=> $ValidatedData['nomorSurat'],
+                'hal'=> $ValidatedData['hal'],
+                'tanggalSurat'=> $ValidatedData['tanggalSurat'],
+                'pemohon'=> $ValidatedData['pemohon'],
+                'tanggalDiTerima'=> $ValidatedData['tanggalDiTerima'],
+                'tiket_id'=> $tiketbaru->id,
+            ]);
+            return redirect('/permohonan');
+        }else{
+            abort(403);
         }
-        $tiket .= $tiket_id; 
-        $tiketbaru = tiket::create([
-            'tiket'=>$tiket,
-            'permohonan'=>1,
-        ]);
-            
-        permohonan::create([
-            'nomorSurat'=> $ValidatedData['nomorSurat'],
-            'hal'=> $ValidatedData['hal'],
-            'tanggalSurat'=> $ValidatedData['tanggalSurat'],
-            'pemohon'=> $ValidatedData['pemohon'],
-            'tanggalDiTerima'=> $ValidatedData['tanggalDiTerima'],
-            'tiket_id'=> $tiketbaru->id,
-        ]);
-        return redirect('/permohonan');
     }
 
     /**
@@ -89,11 +96,13 @@ class PermohonanController extends Controller
      */
     public function show(permohonan $permohonan)
     {
-        //
-        return view('pindaiBarang',[
-            'data'=>$permohonan,
-            
-        ]);
+        if (auth()->user()->jabatan === '01' || auth()->user()->jabatan === '03' || auth()->user()->jabatan === '12') {
+            return view('pindai.Barang',[
+                'data'=>$permohonan,
+            ]);
+        }else{
+            abort(403);
+        }
     }
 
     /**
@@ -118,15 +127,19 @@ class PermohonanController extends Controller
      */
     public function update(Request $request, permohonan $permohonan)
     {
-        $i=0;   
-        foreach ($request->barang as $key) {
-            barang::where('id', $key)->update([
-                'laporan_penilaian_id'=>$request->laporan_penilaian_id,
-                'nilaiWajar'=>$request->value[$i]
-            ]);
-            $i++;
+        if (auth()->user()->jabatan === '01' || auth()->user()->jabatan === '02' || auth()->user()->jabatan === '09' || auth()->user()->jabatan === '10' || auth()->user()->jabatan === '11') {
+            $i=0;   
+            foreach ($request->barang as $key) {
+                barang::where('id', $key)->update([
+                    'laporan_penilaian_id'=>$request->laporan_penilaian_id,
+                    'nilaiWajar'=>$request->value[$i]
+                ]);
+                $i++;
+            }
+            return redirect('/penilaian/'. $permohonan->permohonanPenilaian->id );
+        }else{
+            abort(403);
         }
-        return redirect('/penilaian/'. $permohonan->permohonanPenilaian->id );
     }
 
     /**
@@ -137,15 +150,18 @@ class PermohonanController extends Controller
      */
     public function destroy(permohonan $permohonan)
     {
-        //
-        if ($permohonan->tiket->permohonan === 1) {
-            $permohonan->tiket->update(['permohonan' => 0]);
-            $barang=$permohonan->barang;
-            foreach($barang as $b){
-                $b->delete();
-            };
-            $permohonan->delete();
-            return redirect('/permohonan');
+        if (auth()->user()->jabatan === '01' || auth()->user()->jabatan === '03' || auth()->user()->jabatan === '12') {
+            if ($permohonan->tiket->permohonan === 1) {
+                $permohonan->tiket->update(['permohonan' => 0]);
+                $barang=$permohonan->barang;
+                foreach($barang as $b){
+                    $b->delete();
+                };
+                $permohonan->delete();
+                return redirect('/permohonan');
+            }else{
+                abort(403);
+            }
         }else{
             abort(403);
         }
